Create Imdb review classification
This is a neural network model that tries to classify imdb reviews as either positive or negative based on the text itself. I used keras built in dataset. I first modified the data to replace unknown words and symbols to be easier read for the model. I used a lot of functions of which I don't fully understand but get the gist of. Some credit goes to an youtube channel called tech with tim, who's recommendations I used.