Create main
This is a networking project I made for a programming course. This program stores network data and gives the network as a tree chart. The premise of the problem is, that a program that stores details of a recruited member and the name of the person who recruited that member. This program can store two pairs of names with command "s" and print  a tree chart of stored network with command "p." There are two additional commands. "C" when given id as parameter prints this person's size of subnetwork i.e how many has this person recruited. Command "d" gives the longest chain of recruitment.